rename: RawDataBlockSerializers -> RawBlockCodecs
